refac: utiliza interface http do spring

+++ src/main/java/br/ufrn/imd/feed/model/Feed.java
@@ -15,7 +15,7 @@ public class Feed {
     private int offset;
     private int limit;
     private int size;
-    private Integer newsPosts;
+    private Long newsPosts;
     private List<PostDto> posts;
     private Long totalSongs;
 }
